requirements.txt
gradio
pandas
numpy
langchain
openai
tqdm
python-dotenv
langchain-community
langchain-openai
langchain-chroma
